Simple changes for ILI_Ada_FontTest4
Opensans24 was showing as OpenSans 18

Added print of current font, name, which can help if there is an
issue with the font being displayed.

In particular OpenSans24 is not properly outputting to the screen.

Also updated the config file as the settings I normally use
had ..._RST but the sektchs are looking for _RESET